added jwt filter functionality

+++ muslim-league-core-api/src/main/java/com/marwane/ml_core_api/service/accounts/UserRepository.java
@@ -7,9 +7,11 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends CrudRepository<User,Integer> {
+public interface UserRepository extends CrudRepository<User,Long> {
 
     Optional<User> findByEmail(String email);
 
     Optional<User> findByUsernameOrEmail(String username, String email);
+
+
 }
